WIP: Optionally use window.fetch for loading modules in the esm-amd-loader
When using the loader inside a WKWebView running against local files we end up with requests
for the 'null' origin, which of course will not work at all. But, in this highly controlled environment
one can use a custom window.fetch implementation that would handle 'file://' URLs correctly to avoid that
issue.

This change allows a `use-fetch` attribute on a `<script type="module">`, which modifies the behavior
of the loader to use window.fetch for getting the actual content, and only execute it by setting the
`text` property of a `<script>` tag.

This is very experimental right now!